Added verification for API Key and Device Key.

+++ Prey Installer.applescript
@@ -8,6 +8,7 @@ on ReplaceText(find, replace, subject)
 	return subject
 end ReplaceText
 global selectedmethod
+global installcheck
 on clicked theObject
 	if (name of theObject = "wsselected") then
 		set selectedmethod to true
@@ -39,11 +40,8 @@ on clicked theObject
 		set apth to ReplaceText(" ", "\\ ", apth)
 		set apth to "/Volumes/" & apth & "Contents/Resources/"
 		do shell script "cd " & apth & "
-		sudo rm -Rf /usr/share/prey
-		sudo mkdir /tmp/prey
-		sudo cp config /tmp/prey/temp_config
-		sudo unzip -u prey.zip -d /usr/share/prey
-		" with administrator privileges
+		mkdir /tmp/prey
+		cp config /tmp/prey/temp_config"
 		set idioma to contents of combo box "idioma" of tab view item "general" of tab view "tabsis" of window "Prey 0.3"
 		if (idioma = "Inglés" or idioma = "English") then
 			set idioma to "en"
@@ -54,11 +52,19 @@ on clicked theObject
 		if (selectedmethod is true) then
 			set wsapikey to contents of text field "apikey" of tab view item "wslogin" of tab view "tabsis" of window "Prey 0.3"
 			set wsdevicekey to contents of text field "devicekey" of tab view item "wslogin" of tab view "tabsis" of window "Prey 0.3"
-			do shell script "
-			cd /tmp/prey
-			sed -i -e \"s/lang='.*'/lang='" & idioma & "'/\" temp_config
-			sed -i -e \"s/api_key='.*'/api_key='" & wsapikey & "'/\" temp_config
-			sed -i -e \"s/device_key='.*'/device_key='" & wsdevicekey & "'/\" temp_config" with administrator privileges
+			set wsaklength to (length of wsapikey)
+			set wsdklength to (length of wsdevicekey)
+			if ((wsaklength is not equal to 12) or (wsdklength is not equal to 6)) then
+				set installcheck to false
+				display dialog "Please provide valid API Key & Device Key"
+			else
+				set installcheck to true
+				do shell script "
+				cd /tmp/prey
+				sed -i -e \"s/lang='.*'/lang='" & idioma & "'/\" temp_config
+				sed -i -e \"s/api_key='.*'/api_key='" & wsapikey & "'/\" temp_config
+				sed -i -e \"s/device_key='.*'/device_key='" & wsdevicekey & "'/\" temp_config" with administrator privileges
+			end if
 		else if (selectedmethod is false) then
 			set email to contents of text field "email" of tab view item "mailconfig" of tab view "tabsis" of window "Prey 0.3"
 			set smtpuser to contents of text field "smtpuser" of tab view item "mailconfig" of tab view "tabsis" of window "Prey 0.3"
@@ -77,15 +83,21 @@ on clicked theObject
 			sed -i -e \"s/smtp_username='.*'/smtp_username='" & smtpuser & "'/\" temp_config
 			sed -i -e \"s/smtp_password='.*'/smtp_password='" & enc_pass & "'/\" temp_config" with administrator privileges
 		end if
-		do shell script "
-		cd /tmp/prey
-		sudo cp temp_config /usr/share/prey/config
-		rm -r /tmp/prey
-		(sudo crontab -l | grep -v prey; echo \"*/" & frecuencia & " * * * * /usr/share/prey/prey.sh > /dev/null \") | sudo crontab -" with administrator privileges
 		(* Instalacion Ended*)
-		tell tab view "tabsis" of window "Prey 0.3"
-			set current tab view item to tab view item "finish"
-		end tell
+		if (installcheck is true) then
+			do shell script "
+			sudo rm -Rf /usr/share/prey
+			sudo unzip -u prey.zip -d /usr/share/prey
+			cd /tmp/prey
+			sudo cp temp_config /usr/share/prey/config
+			rm -r /tmp/prey
+			(sudo crontab -l | grep -v prey; echo \"*/" & frecuencia & " * * * * /usr/share/prey/prey.sh > /dev/null \") | sudo crontab -" with administrator privileges
+			tell tab view "tabsis" of window "Prey 0.3"
+				set current tab view item to tab view item "finish"
+			end tell
+		else
+			set installcheck to true
+		end if
 	else if (name of theObject = "quitprey") then
 		tell current application to quit
 	end if
